HomeServer settings recovery: align scheduling connector user keys

+++ includes/homeserver-scheduling-connector-v620.php
@@ -6,7 +6,7 @@ const VP3_HOMESERVER_SCHEDULING_CONNECTOR_V620='homeserver-scheduling-connector-
 function homeserver_scheduling_v620_ensure_schema(PDO $pdo): void
 {
     $pdo->exec("CREATE TABLE IF NOT EXISTS homeserver_scheduling_tokens (
-        user_id BIGINT UNSIGNED NOT NULL PRIMARY KEY,
+        user_id INT UNSIGNED NOT NULL PRIMARY KEY,
         token_hash CHAR(64) NOT NULL UNIQUE,
         token_enc TEXT NOT NULL,
         token_prefix VARCHAR(16) NOT NULL DEFAULT '',
@@ -19,7 +19,7 @@ function homeserver_scheduling_v620_ensure_schema(PDO $pdo): void
     ) ENGINE=InnoDB DEFAULT CHARSET=utf8mb4 COLLATE=utf8mb4_unicode_ci");
     $pdo->exec("CREATE TABLE IF NOT EXISTS homeserver_scheduling_idempotency (
         id BIGINT UNSIGNED NOT NULL AUTO_INCREMENT PRIMARY KEY,
-        user_id BIGINT UNSIGNED NOT NULL,
+        user_id INT UNSIGNED NOT NULL,
         operation VARCHAR(64) NOT NULL,
         idempotency_key VARCHAR(160) NOT NULL,
         response_json LONGTEXT NULL,
